refactor: Unauthorized Screen Component
